25.9.26
重构netGameManager
