+++ Assets/Ri/Scripts/GameManager/NewNetGameManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cb8c3c5600c03bf4aa40835c01d0596a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
